requirements.txt
langchain==0.3.27
langchain-core==0.3.75
langchain-groq==0.2.2
langsmith==0.4.28
python-dotenv==1.1.1
